Move auth and formapp into separate folder
These could be exrracted at some point if desired
