+++ package.json
@@ -158,9 +158,7 @@
     "exclude": [
       "src/**/*.cy.js",
       "src/js/base/**/*.js",
-      "src/js/auth.js",
       "src/js/auth/*.js",
-      "src/js/formapp.js",
       "src/js/formapp/**/*.js",
       "src/js/index.js",
       "src/js/datadog.js",
